rename chatglm to chatglm2 (#4484)
