+++ tests/kit/model_zoo/transformers/chatglm2.py
